# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Lobsense IoT V2 API Documentation",
        "description": "Dokumentasi resmi API sistem monitoring, telemetri sensor tambak, pengontrolan aerator/feeder, inferensi AI, dan manajemen perangkat Lobsense V2.",
        "version": "2.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Lobsense API Local Server"
        }
    ],
    "paths": {
...
        "/api/v2/cages": {
            "get": {
                "tags": [
                    "Cages (KJA)"
                ],
                "summary": "Daftar Semua Keramba (KJA)",
                "description": "Mengambil seluruh data keramba jaring apung yang terdaftar dalam sistem.",
                "operationId": "59868a9d4c911311e4d0013551549391",
                "responses": {
                    "200": {
                        "description": "Daftar keramba berhasil diambil"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Cages (KJA)"
                ],
                "summary": "Tambah Keramba Baru",
                "description": "Membuat data keramba jaring apung (KJA) baru.",
                "operationId": "cad60a320efa4de9be8ec13516f0936a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cage_code"
                                ],
                                "properties": {
                                    "cage_code": {
                                        "type": "string",
                                        "example": "CAGE-B02"
                                    },
                                    "edge_gateway_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": -8.123
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 115.456
                                    },
                                    "volume_cubic_meters": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 12.5
                                    },
                                    "structure_condition": {
                                        "type": "string",
                                        "example": "Baik"
                                    },
                                    "lobster_count": {
                                        "type": "integer",
                                        "example": 250
                                    },
                                    "lobster_age_days": {
                                        "type": "integer",
                                        "example": 45
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Keramba berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    },
                    "401": {
                        "description": "Tidak terautentikasi"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v2/cages/{id}": {
            "get": {
                "tags": [
                    "Cages (KJA)"
                ],
                "summary": "Detail Keramba",
                "description": "Mengambil detail data satu keramba berdasarkan ID.",
                "operationId": "20faff927168460dadc8926121f9b1bf",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail keramba berhasil diambil"
                    },
                    "404": {
                        "description": "Keramba tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Cages (KJA)"
                ],
                "summary": "Update Keramba",
                "description": "Memperbarui data keramba jaring apung yang sudah ada.",
                "operationId": "0764e35856d82cbc729aad6f29be5a76",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cage_code": {
                                        "type": "string",
                                        "example": "CAGE-B02"
                                    },
                                    "edge_gateway_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "latitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": -8.123
                                    },
                                    "longitude": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 115.456
                                    },
                                    "volume_cubic_meters": {
                                        "type": "number",
                                        "format": "double",
                                        "example": 12.5
                                    },
                                    "structure_condition": {
                                        "type": "string",
                                        "example": "Baik"
                                    },
                                    "lobster_count": {
                                        "type": "integer",
                                        "example": 300
                                    },
                                    "lobster_age_days": {
                                        "type": "integer",
                                        "example": 60
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Keramba berhasil diperbarui"
                    },
                    "404": {
                        "description": "Keramba tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cages (KJA)"
                ],
                "summary": "Hapus Keramba",
                "description": "Menghapus data keramba jaring apung berdasarkan ID.",
                "operationId": "11f58c364c8c69dc044496c099ecf461",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Keramba berhasil dihapus"
                    },
                    "404": {
                        "description": "Keramba tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}